arch/x86: Disable GHCB page before starting the decompressed kernel
Currently we are not invoking the sev_es_shutdown_ghcb workflow before
jumping to the decompressed kernel. This is highly required because the
previous GHCB page is still lingering around with page property set as
shared. When that page is allocated by kernel to some other component it
would be in shared state but it was accessed with C-bit set causing the
guest to crash.

As part of the shutdown GHCB workflow we are moving the GHCB page to
encrypted state but hypervisor does not like it because it is registered
as GHCB page to the hypervisor. And hypervisor expects GHCB to be in
shared state. Thus it would fail the pvalidate operation. To fix this
issue temporarily set the GHCB page as 0. This will unset the previously
used GHCB page and guest can function correctly.

Signed-off-by: Jinank Jain <jinankjain@microsoft.com>